requirements.txt
streamlit>=1.46
pymupdf>=1.24
pdfplumber>=0.11
pillow>=10.4
pytesseract>=0.3.10
scikit-learn>=1.5
google-genai>=1.0.0
pydantic>=2.8
python-dotenv>=1.0
